Form submission now works for uploading a book, and related meta data.
TODO: Need to cleanup display page. Correct the navigation menu to include
login, register, and logout buttons, showing them in their correct states.

Would like to be able associate a book with the user that uploaded it, and allow
the person to edit/delete book data. Form validation would also be useful.